Workflow name can now be passed in as an argument.

+++ get_nodes_for_update.sh
@@ -1,7 +1,16 @@
 #!/bin/bash
 WORKFLOW_VERSION=$1
-#TODO: If This value should came from user input, it could make the script mode flexible.
-WORKFLOW_NAME=Workflow_Bundle_
+USER_WORKFLOW_NAME=$2
+# The default workflow name is for Sanger, but this can be overridden by passing some other string as the second argument,
+# for example:
+# get_nodes_for_update.sh 1.0-SNAPSHOT HelloWorld 
+# will select nodes with the HelloWorld bundle, version 1.0-SNAPSHOT
+WORKFLOW_NAME=Workflow_Bundle_SangerPancancerCgpCnIndelSnvStr
+
+if [  ! -z $USER_WORKFLOW_NAME ] ; then
+  WORKFLOW_NAME=Workflow_Bundle_$USER_WORKFLOW_NAME
+fi
+
 # Probably several ways to get a list of worker nodes, but maybe best to just ask vagrant directly.
 # Get a list of everything that varant is managing (that's also in the Bindle directory)
 # $NODES will be an array of directories of vagrant-managed workers.
@@ -9,13 +18,14 @@ LOG_FILE=$(basename $BASH_SOURCE).log
 echo "[ Begin Log, "$(date)" ]">>$LOG_FILE
 {
 echo "Getting list of workers from Vagrant"
-# NOTE: Eventually switch to Adam's inventory script
+
 NODES=($(vagrant global-status | grep ".*running .*/Bindle/.*" | sed 's/.* \(\/home\/.*\/Bindle\/[^\/]*\).*/\1/g'))
 
 if [ ${#NODES[@]} = 0 ] ; then
   echo "Vagrant is not aware of any running nodes. Exiting..."
   exit 1
 fi
+
 
 # Write an inventory file containing all nodes that need to be checked to see if they can be updated.
 #echo "[master]" > candidate_nodes_for_update_inventory
@@ -51,15 +61,16 @@ do
   ANSIBLE_RESULT=$(ansible host_$i -v -i ~/architecture-setup/candidate_nodes_for_update_inventory  -m script -a check-seqware-sanger-last-workflow-status.sh)
   ANSIBLE_STD_OUT=$(echo $ANSIBLE_RESULT |  sed 's/.*"stdout": "\([^"]*\)".*/\1/g')
   ANSIBLE_STD_ERR=$(echo $ANSIBLE_RESULT |  sed 's/.*"stderr": "\([^"]*\)".*/\1/g')
-  #if [[ "$ANSIBLE_STD_OUT" =~ .*The\ status\ of\ the\ last\ job\ is\ completed.* ]] ; then 
-  #  echo "Worker has completed and will be added to the inventory for nodes that will be updated."
+  if [[ "$ANSIBLE_STD_OUT" =~ .*The\ status\ of\ the\ last\ job\ is\ completed.* ]] ; then 
+    echo "Worker has completed and will be added to the inventory for nodes that will be updated."
     NODES_TO_UPDATE+=("$HOST_LINE")
-  #else
-  #  echo "Worker has not completed, message received by Ansible: $ANSIBLE_STD_OUT"
-  #  if [ ! "$ANSIBLE_STD_ERR" = '' ] ; then 
-  #    echo "Std Err: $ANSIBLE_STD_ERR"
-  #  fi
-  #fi
+  else
+    echo "Worker has not completed, message received by Ansible: $ANSIBLE_STD_OUT"
+    if [ ! "$ANSIBLE_STD_ERR" = '' ] ; then 
+      echo "Std Err: $ANSIBLE_STD_ERR"
+    fi
+  fi
+  printf "\n";
 done
 
 if [ ${#NODES_TO_UPDATE[@]} -gt 0 ] ; then
@@ -70,7 +81,7 @@ if [ ${#NODES_TO_UPDATE[@]} -gt 0 ] ; then
   done
   # Now back up the old inventory file, and copy the new one in.
   DATE_STRING=$(date +"%Y%m%d_%H%M%S")
-  cp ~/architecture-setup/pancancer-bag/workflow-update/inventory  ~/architecture-setup/pancancer-bag/workflow-update/inventory_${DATE_STRING}.bkup
+  [ -e ~/architecture-setup/pancancer-bag/workflow-update/inventory ] && cp ~/architecture-setup/pancancer-bag/workflow-update/inventory  ~/architecture-setup/pancancer-bag/workflow-update/inventory_${DATE_STRING}.bkup
   cp nodes_to_update_inventory ~/architecture-setup/pancancer-bag/workflow-update/inventory
 fi
 } | tee -a $LOG_FILE
